feat: chatroom Index.tsx 코드와  동일하게 변경, 파일명 변경
